Implement bracketed lists
Adding is-bracketed() built-in function

Adding $bracketed parameter to join() built-in function

# Conflicts:
#	src/functions.cpp

Modifying emitter to take brackets into account.

First, instead of directly writting the left or right parenthesis
characters, now functions are called for left or right delimiters
and change in case of parenthesis or brackets.

Second, in case of bracketed list, always add the brackets. Also add
comma in case it is a one-item comma-separate list. Otherwise it would
become a one-item space-separated list.

Third, libsass prevent from emitting invisible items. Thus, an empty
space / comma separated list won't be printed at all but bracketed
list are never invisible, event the empty ones.

# Conflicts:
#	src/inspect.cpp

Copying the delimiter property also.

Functions operating on a copy of lists must copy the delimiter also.

# Conflicts:
#	src/functions.cpp

Change the parser to recognize bracketed lists.

This is the idea: in the same way the old parser looks for parenthesis
delimiters, let's add another case for brackets delimiters.

In case an open bracket is found, the parser will look for a list. There
are three possible results of parsing a list:

First, the parsed expression is not a list at all. libsass parser tends
to avoid wrapping if lists are only one item lists. But bracketed lists
prevent unwrapping so the item becomes the solely item for a new
bracketed list.

Example:
singleton-item --> [ singleton-item ]

Second, the list is indeed a list with some explicit delimiter. This
case is similar to the previous one. Brackets prevent unwrapping so this
time is the list as a whole which become the unique item of the
bracketed list.

Example:
(item1, item2, ...) --> [ (item1, item2, ...) ]
[item1, item2, ...] --> [ [item1, item2, ...] ]

Third, the list is a list with no delimiters. So its items actually
belong to the bracketed list. The parse will fix the delimiter of the
list for them to be brackets.

Example:
Suppose | ... | stands for a list with no delimiters.
|item1, item2, ...|  --> [item1, item2, ...]

# Conflicts:
#	src/parser.cpp

Change list definitions to explicitely include delimiters

# Conflicts:
#	src/ast.hpp