Fix import parsing and pass `is_root` around
Imports may or may not be root blocks. Parsing of imports
is done without any knowledge of the parent. Therefore
we must pass the `is_root` state from the actual import
into the parsing function. This reveals a conceptual
problem, since the same import could be once on the
root block and also wrapped inside a selector rule.