Fix win UNC path handling for dot and dotdot directories
This implements a naive path cleanup for windows to resolve
dot and dotdot directories in otherwise already absolute
paths, since UNC file paths must be fully qualified and do
not support dot and dotdot references.

Note that this does collapse x/../y sections into y.
This is by design and wrong, e.g. if /foo on your system is
a symlink to /bar/baz, then /foo/../cd is actually /bar/cd,
not /cd as a naive ../ removal will give you.